refactor unit tests to simplify and fix failures under postgres
