--- tests/Fixture/CustomMetaData/versionsFixture.php
@@ -1,61 +1,0 @@
-<?php
-/**
- * CakePHP(tm) Tests <http://book.cakephp.org/2.0/en/development/testing.html>
- * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          http://book.cakephp.org/2.0/en/development/testing.html CakePHP(tm) Tests
- * @since         1.2.0
- * @license       http://www.opensource.org/licenses/mit-license.php MIT License
- */
-namespace Josegonzalez\Version\Test\Fixture\CustomMetaData;
-
-use Cake\TestSuite\Fixture\TestFixture;
-
-/**
- * Class VersionsFixture
- *
- */
-class VersionsFixture extends TestFixture
-{
-    /**
-     * table property
-     *
-     * @var string
-     */
-    public $table = 'version';
-
-    /**
-     * fields property
-     *
-     * @var array
-     */
-    public $fields = [
-        'id' => ['type' => 'integer'],
-        'version_id' => ['type' => 'integer'],
-        'model' => ['type' => 'string', 'null' => false],
-        'foreign_key' => ['type' => 'integer', 'null' => false],
-        'field' => ['type' => 'string', 'null' => false],
-        'content' => ['type' => 'text'],
-        'custom_field' => ['type' => 'text'],
-        '_constraints' => [
-            'primary' => ['type' => 'primary', 'columns' => ['id']],
-        ],
-    ];
-
-    /**
-     * records property
-     *
-     * @var array
-     */
-    public $records = [
-        ['version_id' => 1, 'model' => 'Articles', 'foreign_key' => 1, 'field' => 'author_id', 'content' => 1, 'custom_field' => 'foo'],
-        ['version_id' => 1, 'model' => 'Articles', 'foreign_key' => 1, 'field' => 'title', 'content' => 'First Article', 'custom_field' => 'foo'],
-        ['version_id' => 1, 'model' => 'Articles', 'foreign_key' => 1, 'field' => 'body', 'content' => 'First Article Body', 'custom_field' => 'foo'],
-        ['version_id' => 1, 'model' => 'Articles', 'foreign_key' => 1, 'field' => 'published', 'content' => 'Y', 'custom_field' => 'foo']
-    ];
-}
